+++ app/charts/bar/bars-state.tsx
@@ -38,6 +38,10 @@ import {
   MOBILE_TOOLTIP_PLACEMENT,
 } from "@/charts/shared/interaction/tooltip-box";
 import { DEFAULT_MARGIN_TOP } from "@/charts/shared/margins";
+import {
+  ShowBandValueLabelsVariables,
+  useShowBandValueLabelsVariables,
+} from "@/charts/shared/show-values-utils";
 import useChartFormatters from "@/charts/shared/use-chart-formatters";
 import { InteractionProvider } from "@/charts/shared/use-interaction";
 import { useSize } from "@/charts/shared/use-size";
@@ -59,7 +63,8 @@ import { ChartProps } from "../shared/ChartProps";
 
 export type BarsState = CommonChartState &
   BarsStateVariables &
-  InteractiveYTimeRangeState & {
+  InteractiveYTimeRangeState &
+  Omit<ShowBandValueLabelsVariables, "offset"> & {
     chartType: "bar";
     xScale: ScaleLinear<number, number>;
     yScaleInteraction: ScaleBand<string>;
@@ -75,7 +80,7 @@ const useBarsState = (
   variables: BarsStateVariables,
   data: ChartStateData
 ): BarsState => {
-  const { chartConfig } = chartProps;
+  const { chartConfig, dimensions, measures } = chartProps;
   const {
     yDimension,
     getX,
@@ -92,6 +97,7 @@ const useBarsState = (
   } = variables;
   const { chartData, scalesData, timeRangeData, paddingData, allData } = data;
   const { fields, interactiveFiltersConfig } = chartConfig;
+  const { x } = fields;
 
   const { width, height } = useSize();
   const formatNumber = useFormatNumber({ decimals: "auto" });
@@ -222,6 +228,13 @@ const useBarsState = (
   };
 
   const barCount = yScale.domain().length;
+  const { offset: xValueLabelsOffset, ...showValuesVariables } =
+    useShowBandValueLabelsVariables(x, {
+      chartData,
+      dimensions,
+      measures,
+      getValue: getX,
+    });
 
   const chartWidth = getChartWidth({ width, left, right });
   const bounds = useChartBounds({ width, chartWidth, height, margins });
@@ -233,7 +246,7 @@ const useBarsState = (
       ? barCount * MIN_BAR_HEIGHT
       : chartHeight;
 
-  xScale.range([0, chartWidth]);
+  xScale.range([0, chartWidth - xValueLabelsOffset]);
   yScaleInteraction.range([0, adjustedChartHeight]);
   yScaleTimeRange.range([0, adjustedChartHeight]);
   yScale.range([0, adjustedChartHeight]);
@@ -309,6 +322,7 @@ const useBarsState = (
     getAnnotationInfo,
     getColorLabel: getSegmentLabel,
     colors,
+    ...showValuesVariables,
     ...variables,
   };
 };